basic audio talking function
this is the first prototype of the listening and the response function and backs it up with chat gpt version 3.5, which is run online as of my current understanding, which is why i have to pay for computing costs. the other downside is, that while listening it is only speaking after creating the whole mp3 file which could be fixed and i'm confident that there are some other efficiencies i can enhance to make the programm faster, for example use a different language porcessing AI as backbone.