removal bug offline v. and video func  online
I cancelled the delete function from the offline Version so it still works, because the videos are in a different folder than the py file.
Got the video function almost working,  it now returns the right link but it still has trouble playing the video in the player for some reason.
removed an unneccessary definiton row