changed some settings for the webhook
i changed the allowed hosts so only the D-ID Server can reach it and created an NGROK URL which is an important variable to store public url created from ngrok.
And I implemented a function into the webhook endpoint so it also responds to get requests cause I'm still not sure wether D-ID sends a get or post request.