Adding websocket + json debug config
I added all the necessary websocket stuff. So the routing and consumers.py and made it so in the webhook function the socket is being used to send the video url to the client without him making a request. Additionally I established a json debug config file. For debugging the websocket. And I made NGROK_URL dynamic.
And I changed updateview because it is not working at all.(But it still doesn't function)