adjustment variables + making frontend post requests work
There were 5 different versions of video id in the code, so I gave them more telling names. I added talk id to the VideoLink Model, so the frontend can better access it. The post request loop from the the frontend wasn't authentificated, so I gave deactivated the csfr authentifaction. I also comented out or removed all the parts of the websocket because I aborted this solution.