ChangeFace Function added + expanded dynamicality + Frontend upgrade + update openai api + requirements.txt update
I made so it is possible to switch the face of the AI Twin. I did that with the SDK API. additionally I removed some hardcoded variables. I also upgraded the frontend with a more natural design and the drag and drop feature for the images is now implemented. Addtitonally openai Api released an update so the way of the usage of it slightly changed. and I updated requirements.txt so its now complete.